+++ schemacrawler-api/src/main/java/schemacrawler/schemacrawler/SchemaRetrievalOptionsBuilder.java
@@ -57,7 +57,7 @@ public final class SchemaRetrievalOptionsBuilder
   EnumDataTypeHelper enumDataTypeHelper;
   Map<SchemaInfoMetadataRetrievalStrategy, MetadataRetrievalStrategy> metadataRetrievalStrategyMap;
   Consumer<Connection> connectionInitializer;
-  HostIdentityExtractor hostIdentityExtractor;
+  HostLocationExtractor hostLocationExtractor;
 
   private SchemaRetrievalOptionsBuilder() {
     dbServerType = DatabaseServerType.UNKNOWN;
@@ -73,7 +73,7 @@ public final class SchemaRetrievalOptionsBuilder
         TableTypes.from("TABLE", "VIEW", "SYSTEM TABLE", "GLOBAL TEMPORARY", "LOCAL TEMPORARY");
     enumDataTypeHelper = NO_OP_ENUM_DATA_TYPE_HELPER;
     connectionInitializer = connection -> {};
-    hostIdentityExtractor = null;
+    hostLocationExtractor = null;
 
     metadataRetrievalStrategyMap = new EnumMap<>(SchemaInfoMetadataRetrievalStrategy.class);
     for (final SchemaInfoMetadataRetrievalStrategy key :
@@ -131,7 +131,7 @@ public final class SchemaRetrievalOptionsBuilder
     overridesTypeMap = Optional.empty();
     metadataRetrievalStrategyMap = options.getMetadataRetrievalStrategyMap();
     connectionInitializer = options.getConnectionInitializer();
-    hostIdentityExtractor = options.getHostIdentityExtractor();
+    hostLocationExtractor = options.getHostLocationExtractor();
 
     return this;
   }
@@ -169,9 +169,9 @@ public final class SchemaRetrievalOptionsBuilder
     return this;
   }
 
-  public SchemaRetrievalOptionsBuilder withHostIdentityExtractor(
-      final HostIdentityExtractor hostIdentityExtractor) {
-    this.hostIdentityExtractor = hostIdentityExtractor;
+  public SchemaRetrievalOptionsBuilder withHostLocationExtractor(
+      final HostLocationExtractor hostLocationExtractor) {
+    this.hostLocationExtractor = hostLocationExtractor;
     return this;
   }